Rename lammps preprocessor code.
