+++ src/diffusion_for_multi_scale_molecular_dynamics/data/diffusion/lammps_processor_for_diffusion.py
