+++ tests/data/diffusion/test_data_preprocess.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from diffusion_for_multi_scale_molecular_dynamics.data.diffusion.data_preprocess import \
+from diffusion_for_multi_scale_molecular_dynamics.data.diffusion.lammps_processor_for_diffusion import \
     LammpsProcessorForDiffusion
 from diffusion_for_multi_scale_molecular_dynamics.namespace import (
     CARTESIAN_FORCES, CARTESIAN_POSITIONS, RELATIVE_COORDINATES)
